fix: 방 생성시 방장 자동 참여, 방장은 참여 취소 불가 처리

+++ backend/billboard/src/main/java/com/ssafy/billboard/model/service/RoomServiceImpl.java
@@ -25,6 +25,7 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     public RoomDto.RoomInfo createRoom(RoomDto.RoomInput roomInput){
+        //없는 유저가 방장 > null
         if(!userRepository.existsByUserId(roomInput.getHostId()))
             return null;
         Room room = roomRepository.save(Room.builder()
@@ -36,6 +37,8 @@ public class RoomServiceImpl implements RoomService {
                 .lng(roomInput.getLng())
                 .date(roomInput.getDate())
                 .build());
+        //방장은 자동 참여 처리
+        createEntry(room.getRoomId(), room.getHostId());
         return RoomDto.RoomInfo.builder()
                 .roomId(room.getRoomId())
                 .hostId(room.getHostId())
@@ -88,11 +91,13 @@ public class RoomServiceImpl implements RoomService {
                     .replies(getReplies(roomId))
                     .build();
         }
+        //없는 방아이디 > null
         return null;
     }
 
     @Override
     public boolean deleteRoom(long roomId){
+        //없는 방 아이디 > false
         if(!roomRepository.existsById(roomId))
             return false;
         roomRepository.deleteById(roomId);
@@ -101,6 +106,7 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     public boolean updateRoom(long roomId, RoomDto.RoomUpdate roomUpdate){
+        //없는 방 아이디 > false
         if(!roomRepository.existsById(roomId))
             return false;
         Room room = roomRepository.findById(roomId).get();
@@ -111,6 +117,7 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     public boolean createReply(long roomId, RoomDto.ReplyInput replyInput){
+        //없는 방 or 유저 > false;
         if(!roomRepository.existsById(roomId) || !userRepository.existsByUserId(replyInput.getUserId()))
             return false;
         replyRepository.save(Reply.builder()
@@ -123,6 +130,7 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     public List<RoomDto.ReplyInfo> getReplies(long roomId){
+        //없는 방 아이디 > null
         if(!roomRepository.existsById(roomId))
             return null;
         List<Reply> repliesEntity = replyRepository.findAllByRoomId(roomId);
@@ -139,6 +147,7 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     public boolean deleteReply(long replyId){
+        //없는 댓글 아이디 > false
         if(!replyRepository.existsById(replyId))
             return false;
         replyRepository.deleteById(replyId);
@@ -147,29 +156,37 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     public int createEntry(long roomId, String userId){
+        //없는 방 or 없는 유저 > 0
         if(!roomRepository.existsById(roomId) || !userRepository.existsByUserId(userId))
             return 0;
+        //해당 방에 해당 유저가 이미 참여중 > -1
         if(entryRepository.existsByRoomAndUserId(roomRepository.findById(roomId).get(), userId))
             return -1;
+        Room room = roomRepository.findById(roomId).get();
         entryRepository.save(Entry.builder()
-                .room(roomRepository.findById(roomId).get())
+                .room(room)
                 .userId(userId)
                 .build());
         return 1;
     }
 
     @Override
-    public boolean deleteEntry(long roomId, String userId){
+    public int deleteEntry(long roomId, String userId){
+        //없는 방 or 유저 or 참여중이 아님 : 0
         if(!roomRepository.existsById(roomId)
                 || !userRepository.existsByUserId(userId)
                 || !entryRepository.existsByRoomAndUserId(roomRepository.findById(roomId).get(), userId))
-            return false;
+            return 0;
+        //방장이 참여 취소 : -1
+        if(roomRepository.findById(roomId).get().getHostId().equals(userId))
+            return -1;
         entryRepository.deleteByRoomAndUserId(roomRepository.findById(roomId).get(), userId);
-        return true;
+        return 1;
     }
 
     @Override
     public List<RoomDto.RoomReservationInfo> getRoomsByUserId(String userId){
+        //없는 유저 > null
         if(!userRepository.existsByUserId(userId))
             return null;
         List<Entry> entries = entryRepository.findAllByUserId(userId);
